+++ .gitignore
@@ -4,3 +4,5 @@ node_modules/
 .DS_Store
 
 images/
+
+project_files/
