+++ README.md
@@ -3,3 +3,5 @@
 By Marshall Protsman, Great Grandson of Tommy Protsman (founder: Stone Mountain Park Auto Museum).
 
 Paper background downloaded from www.subtlepatterns.com
+
+favicon by Bohdan Burmich from Noun Project
