--- src/components/Landing/Landing.css
@@ -1,23 +1,0 @@
-/* .Sandbox {
-    background-color: rgb(19, 20, 20);
-} */
-
-/* h1 {
-    color: rgb(243, 245, 252);
-} */
-
-/* .ContentStyle {
-  height: 80vh;
-  width: 92vw;
-  margin: 20px auto;
-  border-style: solid;
-  border-width: 1px;
-  padding-top: 33vh;
-} */
-
-/* 
-.body-items {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-} */
