+++ src/images.js
@@ -1,18 +1,20 @@
-import Tommy from "./images/TommyAndElizabethComp1.png"
-import TommyR from "./images/TommyAndElizabeth-Reverse.png"
-import Front from "./images/FrontComp1.png"
+// import Tommy from "./images/TommyAndElizabethComp1.png"
+// import TommyR from "./images/TommyAndElizabeth-Reverse.png"
+// import Front from "./images/FrontComp1.png"
+// import FrontR from "./images/Front-Reverse.png"
 
 const Photos = [
     {
-    image_src: Tommy,
-    image_reverse: TommyR,
+    image_src: "https://res.cloudinary.com/dxf0lgyn4/image/upload/v1587406779/Postcards/TommyAndElizabethComp1_abbu9b.png",
+    image_reverse: "https://res.cloudinary.com/dxf0lgyn4/image/upload/v1587406771/Postcards/TommyAndElizabeth-Reverse_xibunu.png",
     title: "Tommy And Elizabeth 1908 Buick",
     alt: "Tommy And Elizabeth 1908 Buick",
     description:
       "A 1908 Buick Model X Roadster, 4 Cylinder, 18 Horse Power. The rear seat was known as the mother-in-law seat."
     },
   {
-    image_src: Front,
+    image_src: "https://res.cloudinary.com/dxf0lgyn4/image/upload/v1587406809/Postcards/FrontComp1_tcy2b9.png",
+    image_reverse: "https://res.cloudinary.com/dxf0lgyn4/image/upload/v1587406756/Postcards/Front-Reverse_um4hph.png",
     title: "Storefront of the auto museum",
     alt: "Storefront of the auto museum",
     description:
