+++ src/index.css
@@ -1,12 +1,24 @@
 body {
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
+  font-family: -apple-system, "Helvetica Neue",
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   background: rgb(255, 246, 220);
   color: rgb(0, 0, 0);
+}
+
+h1 {
+font-family: -apple-system, "Helvetica Neue", sans-serif;
+}
+
+h4 {
+  font-family: -apple-system, Lora, Karla-Regular, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+}
+
+p {
+  font-family: -apple-system, Lora, Karla-Regular, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+
 }
 
 code {
@@ -24,13 +36,24 @@ li {
 
 
 /* INDIVIDUAL COMPONENT STYLING */
+.title {
+display: flex;
+}
 
+.main-title {
+  display: flex;
+  justify-content: left;
+  font-family: -apple-system, "Helvetica Neue", sans-serif;
+  font-size: 150px;
+  color: #DC2C21;
+  line-height: 130px;
+}
 
 /* Individual sections */
 .home-section {
-  background-color: #E3D0AC;
+  background-color: #E6DFD0;
   display: flex;
-  justify-content: center;
+  /* justify-content: left; */
 }
 
 .section1 {
@@ -52,6 +75,8 @@ li {
   flex-direction: column;
 }
 
+/* FLEXING */
+
 .flex-center {
   display: flex;
   justify-content: center;
@@ -62,6 +87,12 @@ li {
   display: flex;
   flex-direction: column-reverse;
 }
+
+.flex-left {
+  display: flex;
+  justify-content: left;
+  margin-left: 10px;
+  }
 
 .btn-clear {
   border: none;
@@ -83,7 +114,8 @@ width: 100vh;
   display: flex;
   flex-direction: column;
   justify-content: space-around;
-  align-items: center;
+  align-items: left;
+  margin: 15px;
 }
 
 .direction {
